refactor: Utils file
