+++ online-store/src/interface/index.ts
@@ -1,0 +1,5 @@
+export interface IUtils {
+  resource: string | null;
+  id: string | null;
+  verb: string | null;
+}
